refactor: share classifyDistinguishedTier from ClubEligibilityUtils (#1124)
/simplify pass — the tier classifier lives next to
DISTINGUISHED_STATUS_CODES/isDistinguishedStatusCode (the consolidation
site from #1095) instead of as a DataTransformer private, exported via
the analytics barrel; calculateTotals folds its three reduce passes
into the single tier-counting loop.
